% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DGLM_norm}
\alias{DGLM_norm}
\title{DGLM-norm}
\usage{
DGLM_norm(m.form, d.form, data, maxiter = 20, conv = 1e-06)
}
\arguments{
\item{m.form}{mean formula}

\item{d.form}{variance formula}

\item{data}{data}

\item{maxiter}{max number of iterations to do}

\item{conv}{converges when LL changes by less than x}
}
\value{
a list of length 3
}
