% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_dglm}
\alias{fit_dglm}
\title{fit_dglm}
\usage{
fit_dglm(mean_alt_formula, var_alt_formula, mean_null_formula, var_null_formula,
  this_locus_df)
}
\arguments{
\item{mean_alt_formula}{yada}

\item{var_alt_formula}{yada}

\item{mean_null_formula}{yada}

\item{var_null_formula}{yada}

\item{this_locus_df}{yada}
}
\value{
a one-row data_frame
}
\description{
fit_dglm
}
