% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomeScan.R
\name{GenomeScan}
\alias{GenomeScan}
\title{GenomeScan}
\usage{
GenomeScan(meanFormula, varFormula, file.names, verbose = FALSE)
}
\arguments{
\item{phenTable, }{a table of phenotypes to be validated}

\item{genTable, }{a table of genotypes to be validated}
}
\value{
object of class c('tbl_df', 'data.frame') with []
}
\author{
Robert Corty
}

