% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDGSNPByName.R
\name{GetDGSNPByName}
\alias{GetDGSNPByName}
\title{GetDGSNPByName Function}
\usage{
GetDGSNPByName(file.names, snp.name)
}
\arguments{
\item{file.names}{a list of 3.}
}
\value{
the SNP in a named vector
}
\description{
This function accesses a snp by name based on a map.file,
a list of genotype.files, and a snp.name
}
\author{
Robert Corty
}

