% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestigateDGSNP.R
\name{InvestigateDGSNP}
\alias{InvestigateDGSNP}
\title{InvestigateDGSNP}
\usage{
InvestigateDGSNP(phenotype.name, snp.name, snp.values = NULL, chr = NULL,
  file.names = NULL, modeling.df = NULL, dot.color = NULL,
  keep.only = NULL, mean.formula = NULL, var.formula = NULL,
  verbose.title = FALSE, draw.violins = TRUE, draw.scatter = TRUE,
  draw.sd.bars = TRUE, draw.summary.stats = TRUE, alpha.scatter = 0.4,
  alpha.violin = 0.4)
}
\arguments{
\item{phenotype.name}{name of the phenotype}

\item{snp.name}{name of the snp}

\item{file.names}{names of phenotype and genotype files}

\item{mean.formula}{the mean formula}

\item{var.formula}{the variance formula}

\item{draw.violins}{should the violion plot be plotted}

\item{draw.scatter}{should the scatter plot be plotted}

\item{alpha.scatter}{alpha on the scatter plot}

\item{alpha.violin}{alpha on the violin}

\item{draw.ns}{should the n per group be plotted}

\item{draw.sd.bar}{should the bars indicating the standard deviation of each genotype be plotted}
}
\value{
nothing, just plots
}
\description{
Model and plot to investigate a directly genotyped single
nucleotide polymorphism
}
\author{
Robert W. Corty
}
