% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHSList.R
\docType{data}
\name{LHSList}
\alias{LHSList}
\title{LHSList}
\format{An object of class \code{result} of length 1.}
\usage{
LHSList
}
\arguments{
\item{value}{}
}
\examples{

}
\author{
Gabor Csardi
}
\keyword{datasets}
