% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\docType{class}
\name{mvGWAS}
\alias{mvGWAS}
\title{A Reference Class to represent a mean-variance GWAS}
\description{
A reference class implementing a Mean-Variance Genome Wide Association Study
}
\section{Fields}{

\describe{
\item{\code{metadata}}{information about the mvGWAS}

\item{\code{data}}{the data used to conduct the mvGWAS}

\item{\code{null_model}}{the model fit without any SNPs}

\item{\code{results}}{the results of the mvGWAS}
}}


