% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_control.R
\name{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\title{apply_genomic_control}
\value{
TRUE if successful, FALSE otherwise.

TRUE if successful, FALSE otherwise.

TRUE if successful, FALSE otherwise.
}
\description{
Applies genomic control

#' @title apply_genomic_control_DS_
#' @name mvGWAS_apply_genomic_control
#'
#' @return TRUE if successful, FALSE otherwise.
#'
#' @title gc_DS_one_test_
#' @name mvGWAS_internals
#'
#' @return TRUE if successful, FALSE otherwise.
#' @importFrom dplyr %>%
#' @importFrom rlang UQ
#' @importFrom rlang sym
#'
}
