% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\name{mvGWAS_conduct_scan}
\alias{mvGWAS_conduct_scan}
\title{conduct_scan}
\arguments{
\item{mean_formula}{the mean formula}

\item{var_formula}{the variance formula}

\item{computer}{Either 'local' (default) or 'slurm'. Should the scan be conducted on the local machine or on a slurm cluster.}
}
\value{
the mvGWAS object  (change this?)
}
\description{
conducts an mvGWAS
}
