% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\name{mvGWAS_initialize}
\alias{mvGWAS_initialize}
\title{initialize}
\arguments{
\item{phenotype_file}{The file that stores the phenotypes.  Must be a .CSV or .RDS.  Must have one column called 'ID'.}

\item{genotype_directory}{The directory that stores the genotypes.  They must be in .vcf.gz format.}
}
\value{
an mvGWAS object
}
\description{
Initialize an mvGWAS object
}
