% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_scan.R, R/genomic_control.R
\name{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\title{scan_vcf_file_}
\arguments{
\item{formula}{the formula to validate}

\item{type}{one of 'mean' or 'var'}
}
\value{
nothing

nothing

nothing

TRUE if successful, FALSE otherwise.

TRUE if successful, FALSE otherwise.
}
\description{
determines which keywords are used in mean_formula and var_formula

validates a formula and stashes it for use in an mvGWAS
}
