% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mvGWAS_manhattan_plot}
\alias{mvGWAS_manhattan_plot}
\title{Manhattan Plot}
\arguments{
\item{what_to_plot}{what to plot}

\item{lowest_nlog_p_to_plot}{lowest neg log10 p to plot}

\item{highlight_nlog_p_above}{highlight neg log10 p value above this number with larger points}

\item{highlight_size}{how big are the highlight points}

\item{nonhighlight_alpha}{how transparent are the non-highlighted points}
}
\value{
the plot
}
\description{
makes a manhattan plot for a mvGWAS
}
