% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutationNullByDGSNP.R
\name{PermutationNullByDGSNP}
\alias{PermutationNullByDGSNP}
\title{PermutationNullByDGSNP Function}
\usage{
PermutationNullByDGSNP(phenotype.name, snp.name, file.names,
  box.cox.lambda = 1)
}
\arguments{
\item{phenotype.name}{the phenotype to plot}

\item{snp.name}{the rsid of the SNP to stratify by}

\item{draw.ns}{should the plot show the number of individuals in each
genotype group?  Defaults to FALSE}
}
\value{
No return.  Just the plot.
}
\description{
This function calculates the
}
\author{
Robert Corty
}

