% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhenotypeByDGSNP.R
\name{PhenotypeByDGSNP}
\alias{PhenotypeByDGSNP}
\title{PhenotypeByDGSNP Function}
\usage{
PhenotypeByDGSNP(phenotype.name, snp.name, file.names, box.cox.lambda = 1,
  mean.formula = NULL, var.formula = NULL, draw.ns = FALSE,
  violin = FALSE, scatter = TRUE, sd.bars = TRUE, alpha.scatter = 0.4,
  alpha.violin = 0.4, ...)
}
\arguments{
\item{phenotype.name}{the phenotype to plot}

\item{snp.name}{the rsid of the SNP to stratify by}

\item{draw.ns}{should the plot show the number of individuals in each
genotype group?  Defaults to FALSE}
}
\value{
No return.  Just the plot.
}
\description{
This function plots the specified phenotype
stratified by the genotype at the specified SNP.
options [a, b, and c] must have been set a priori.
}
\author{
Robert Corty
}

