% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQPlot.R
\name{QQPlot}
\alias{QQPlot}
\title{QQPlot Function}
\usage{
QQPlot(v, theoretical = "unif", do.neg.log.ten = TRUE, color = "black")
}
\arguments{
\item{v}{values to be plotted}

\item{theoretical}{theoretical distribution of the values \code{v},
defaults to 'unif'.  Other possible values are 'normal'}

\item{plot.neg.log.ten}{Whether \code{v} and its theoretical values should be
negative-log-tenned before plotting.}
}
\value{
No return.  Just the plot.
}
\description{
This function takes a vector of values, sorts them, and plots them
against the expected distribution specified by the \code{theoretical} argument.
As \code{v} gets longer, the less interesting end(s) of the distribution get
sparisifed to save plotting time.
}
\details{
For uniform theoretical distribution, the 'less interesting' end
is considered the end near 1.  The values are always sorted at the start.
The 'most interesting' 1e3 values are always plotted.  The values up through
1e4 are plotted in a '1 out of every 10' fashion.  The values up through 1e5
are plotted in a '1 out of every 100 fashion'.  The values up through 1e6 are
plotted in a '1 out of every 1000 fashion'.  The values up through 1e7 are
plotted in a '1 out of every 10000' fashion.
}
\author{
Robert Corty
}
