% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryNA}
\alias{tryNA}
\title{tryNA}
\usage{
tryNA(expr)
}
\arguments{
\item{expr}{the expression to try to evaluate}
}
\value{
If the expression evaluates without error, the evaluated expression.
If there is an error in evaluating the expression, NA.
Warnings are suppresssed.
}
\description{
silently get what you can out of an expression
}
