% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tryNULL}
\alias{tryNULL}
\title{tryNULL}
\usage{
tryNULL(expr)
}
\arguments{
\item{expr}{the expression to try to evaluate}
}
\value{
If the expression evaluates without error, the evaluated expression.
If there is an error in evaluating the expression, NULL.
Warnings are suppresssed.
}
\description{
silently get what you can out of an expression
}
