% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ValidateGenotypes.R
\name{ValidateGenotypes}
\alias{ValidateGenotypes}
\title{ValidateGenotypes}
\usage{
ValidateGenotypes(genTable)
}
\arguments{
\item{genTable,}{a table of genotypes to be validated}
}
\value{
TRUE if genTable is a valid genotype table for GWASviaDGLM, FALSE otherwise
}
\description{
ValidateGenotypes
}
\author{
Robert Corty
}

