% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidatePhenotypes.R
\name{ValidatePhenotypes}
\alias{ValidatePhenotypes}
\title{ValidatePhenotypes}
\usage{
ValidatePhenotypes(phenTable)
}
\arguments{
\item{phenTable, }{a table of phenotypes to be validated}
}
\value{
TRUE if phenTable is a valid phenotype table for GWASviaDGLM, FALSE otherwise
}
\author{
Robert Corty
}

