% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateSNPValues.R
\name{ValidateSNPValues}
\alias{ValidateSNPValues}
\title{ValidateSNPValues}
\usage{
ValidateSNPValues(snp.values, phen.df)
}
\arguments{
\item{snp.values}{the snp values}

\item{phen.df}{the phenotype data.frame}
}
\value{
nothing if the snp values are valid, errors if they're not
}
\author{
Robert W. Corty
}
