% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoValidatePhenotypesAndGenotypes.R
\name{CoValidatePhenotypesAndGenotypes}
\alias{CoValidatePhenotypesAndGenotypes}
\title{CoValidatePhenotypesAndGenotypes}
\usage{
CoValidatePhenotypesAndGenotypes(phenTable, genTable)
}
\arguments{
\item{phenTable, }{a table of phenotypes to be validated}

\item{genTable, }{a table of genotypes to be validated}
}
\value{
TRUE if phenTable and genTable are co-valid for use in GWASviaDGLM
}
\author{
Robert Corty
}
