% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\docType{class}
\name{mvGWAS}
\alias{mvGWAS}
\title{mvGWAS reference class}
\description{
A reference class implementing a Mean-Variance Genome Wide Association Study
}
\section{Fields}{

\describe{
\item{\code{metadata}}{information about the mvGWAS}

\item{\code{data}}{the data used to conduct the mvGWAS}

\item{\code{intermediates}}{computations stored in the course of conducting the mvGWAS that aren't readily interpretable}

\item{\code{results}}{the results of the mvGWAS}
}}


