% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\name{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\alias{mvGWAS_apply_genomic_control}
\title{apply_genomic_control}
\value{
TRUE if successful, FALSE otherwise.

TRUE if successful, FALSE otherwise.

TRUE if successful, FALSE otherwise.
}
\description{
Applies genomic control
}
