% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_scan.R
\name{mvGWAS_conduct_scan}
\alias{mvGWAS_conduct_scan}
\alias{mvGWAS_conduct_scan}
\alias{mvGWAS_conduct_scan}
\title{conduct_scan}
\arguments{
\item{mean_formula}{the mean formula}

\item{var_formula}{the variance formula}

\item{computer}{Either 'local' (default) or 'slurm'. Should the scan be conducted on the local machine or on a slurm cluster.}

\item{num_cores}{the number of cores to use, defaults to parallel::detectCores()}
}
\value{
the mvGWAS object  (change this?)

nothing

nothing
}
\description{
conducts an mvGWAS

conducts a genome scan locally

conducts a genome scan on a slurm cluster
}
