% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R
\name{mvGWAS_internals_}
\alias{mvGWAS_internals_}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\title{scan_vcf_file_}
\arguments{
\item{formula}{the formula to validate}

\item{type}{one of 'mean' or 'var'}

\item{num_cores}{the number of cores to use, defaults to parallel::detectCores()}
}
\value{
nothing

nothing

nothing

nothing
}
\description{
validates a formula and stashes it for use in an mvGWAS

conducts a genome scan locally

conducts a genome scan on a slurm cluster
}
