% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGWAS.R, R/utils.R
\name{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals}
\alias{mvGWAS_internals_}
\alias{scan_vcf_file_}
\title{stash_formula_}
\usage{
scan_vcf_file_(file_name, drop_gts_w_fewer_than_x_obs = 5)
}
\arguments{
\item{formula}{the formula to validate}

\item{type}{one of 'mean' or 'var'}

\item{num_cores}{the number of cores to use, defaults to parallel::detectCores()}
}
\value{
nothing

nothing

nothing

nothing

nothing
}
\description{
validates a formula and stashes it for use in an mvGWAS

determines which keywords are used in mean_formula and var_formula

conducts a genome scan locally

conducts a genome scan on a slurm cluster
}
